src/modules/module_24_prescription_audit/backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
pymongo==4.9.0
python-dotenv==1.0.1
marshmallow==3.23.0
bcrypt==4.2.0
PyJWT==2.10.0
